+++ ChangeLog.rst
@@ -5,13 +5,21 @@ Changes
 #######
 
 Develop
-==========================
+=======
 
-- #1: Add missing license (GPL).
+Incompatible Changes
+--------------------
+
 - Change to CMake as build system.
   See :ref:`installation` for how to build the package,
   or use it as a submodule in a nother CMake project.
 - Require version v0.4 PermGroup, though it is available as a Git submodule.
+- Change :cpp:func:`make_default_visitor` to include many more visitors,
+  to provide a quick way to get a reasonable algorithm configuration.
+
+New Features
+------------
+
 - Add the program :ref:`graph_canon_run` for easier invocation
   of canonicalization programs on batches of graphs.
 - Add script for downloading various graph collections:
@@ -19,10 +27,20 @@ Develop
 - Add the scripts :ref:`graph_canon_dreadnaut` and :ref:`graph_canon_bliss`
   for providing a common interface between the programs in this package,
   Bliss, nauty, and Traces.
+- Add ``num_edges`` support for :cpp:class:`ordered_graph`.
+- Add :cpp:func:`as_range` helper function for converting Boost.Graph
+  iterator pairs to ranges.
+
+Other
+-----
+
+- #1: Add missing license (GPL).
+- Add and update examples.
+- Lots of documentation improvements.
 
 
 Release v0.3 (2018-09-04)
-==========================
+=========================
 
 - Require version v0.3 PermGroup.
 - Add Schreier-Sims-based automorphism pruning.
@@ -33,6 +51,6 @@ Release v0.3 (2018-09-04)
 
 
 Release v0.2 (2018-02-05)
-==========================
+=========================
 
 Second public version.